requirements.txt
altair==5.4.1
anyio==4.4.0
archspec==0.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
boltons==24.0.0
branca==0.8.0
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
conda==24.7.1
conda-libmamba-solver==24.7.0
conda-package-handling==2.3.0
conda-package-streaming==0.10.0
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
folium==0.18.0
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
h2==4.1.0
haversine==2.8.1
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.8
importlib-metadata==8.4.0
importlib-resources==6.4.4
ipykernel==6.29.5
ipympl==0.9.4
ipython==8.26.0
ipython_genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
libmambapy==1.5.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst==2.1.2
mistune==3.0.2
munkres==1.1.4
narwhals==1.13.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
notebook_shim==0.2.4
numpy==2.1.0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pip==24.3.1
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.24.1
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.28.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==18.0.0
pycosat==0.6.6
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.14.1
Send2Trash==1.8.3
setuptools==72.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.2
streamlit==1.40.0
streamlit_folium==0.23.1
tenacity==9.0.0
terminado==0.18.1
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
truststore==0.9.2
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
typing-utils==0.1.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
watchdog==5.0.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
widgetsnbextension==4.0.13
win-inet-pton==1.1.0
xyzservices==2024.9.0
zipp==3.20.1
zstandard==0.23.0